chore: restructure the package to make it more reasonable by splitting webview and app
