+++ app/src/main/java/com/superproductivity/superproductivity/app/SpTask.kt
@@ -1,4 +1,4 @@
-package com.superproductivity.superproductivity
+package com.superproductivity.superproductivity.app
 
 data class SpTask(
     val id: String,
